Use strict error handling in bash scripts (#3083)

+++ upload_output.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+set -xe -o pipefail -o nounset
+
 BRANCH=$BRANCH_NAME
 ENDPOINT_URL=$S3_ENDPOINT_URL
 S3_PATH=$S3_UPLOAD_PATH
@@ -13,18 +15,25 @@ aws s3 sync output/json/_build_en/json/_images $S3_PATH/$BRANCH/images_en --endp
 aws s3 sync output/json/_build_ru/json/_images $S3_PATH/$BRANCH/images_ru --endpoint-url=$ENDPOINT_URL --delete --size-only
 
 # upload pdf files
-aws s3 cp --acl public-read output/_latex_en/Tarantool.pdf $S3_PATH/$BRANCH/Tarantool-en.pdf --endpoint-url=$ENDPOINT_URL
-aws s3 cp --acl public-read output/_latex_ru/Tarantool.pdf $S3_PATH/$BRANCH/Tarantool-ru.pdf --endpoint-url=$ENDPOINT_URL
+if [ -f output/_latex_en/Tarantool.pdf ]; then
+  aws s3 cp --acl public-read output/_latex_en/Tarantool.pdf $S3_PATH/$BRANCH/Tarantool-en.pdf --endpoint-url=$ENDPOINT_URL
+fi
+if [ -f output/_latex_ru/Tarantool.pdf ]; then
+  aws s3 cp --acl public-read output/_latex_ru/Tarantool.pdf $S3_PATH/$BRANCH/Tarantool-ru.pdf --endpoint-url=$ENDPOINT_URL
+fi
 
 # upload singlehtml and assets
+if [ -f output/html/en/singlehtml.html ]; then
 aws s3 sync --acl public-read output/html/en/_static $S3_PATH/$BRANCH/en/_static --endpoint-url=$ENDPOINT_URL --delete --size-only
 aws s3 sync --acl public-read output/html/en/_images $S3_PATH/$BRANCH/en/_images --endpoint-url=$ENDPOINT_URL --delete --size-only
 aws s3 cp --acl public-read output/html/en/singlehtml.html $S3_PATH/$BRANCH/en/singlehtml.html --endpoint-url=$ENDPOINT_URL
+fi
+if [ -f output/html/ru/singlehtml.html ]; then
 aws s3 sync --acl public-read output/html/ru/_static $S3_PATH/$BRANCH/ru/_static --endpoint-url=$ENDPOINT_URL --delete --size-only
 aws s3 sync --acl public-read output/html/ru/_images $S3_PATH/$BRANCH/ru/_images --endpoint-url=$ENDPOINT_URL --delete --size-only
 aws s3 cp --acl public-read output/html/ru/singlehtml.html $S3_PATH/$BRANCH/ru/singlehtml.html --endpoint-url=$ENDPOINT_URL
+fi
 
-set -xe
 curl --fail --show-error \
     --data '{"update_key":"'"$TARANTOOL_UPDATE_KEY"'"}' \
     --header "Content-Type: application/json" \
